@@ -1,20 +1,28 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net8.0-windows</TargetFramework>
+    <TargetFramework>net10.0-windows</TargetFramework>
     <OutputType>Library</OutputType>
     <PlatformTarget>x64</PlatformTarget>
     <Platforms>x64;AnyCPU</Platforms>
-    <AssemblyVersion>1.1</AssemblyVersion>
+    <AssemblyVersion>1.2</AssemblyVersion>
   </PropertyGroup>
+  
+  <PropertyGroup>
+    <DalamudLibPath>$(appdata)\XIVLauncher\addon\Hooks\dev\</DalamudLibPath>
+  </PropertyGroup>
+  
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
   </PropertyGroup>
+  
+  <ItemGroup>
+    <PackageReference Include="DalamudPackager" Version="14.0.1" />
+  </ItemGroup>
+  
   <ItemGroup>
     <Reference Include="Dalamud">
-      <HintPath>..\Dalamud\Dalamud.dll</HintPath>
-      <Private>False</Private>
+      <HintPath>$(DalamudLibPath)Dalamud.dll</HintPath>
+      <Private>false</Private>
     </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="DalamudPackager" Version="13.0.0" />
   </ItemGroup>
 </Project>
